requirements.txt
asyncio==3.4.3            # Стандартная библиотека Python для работы с асинхронным программированием
aiogram>=3.8.0            # Библиотека для создания ботов на платформе Telegram с использованием async/await
aiogram-dialog==2.1.0     # Библиотека для управления диалогами в Telegram-ботах на базе aiogram

cachetools==5.5.0         # Библиотека для кэширования данных в Python
environs==11.2.1          # Библиотека для управления конфигурациями и переменными окружения
python-dotenv>=1.0.1      # Библиотека для загрузки переменных окружения из файла .env
icecream==2.1.3           # Библиотека для дебагинга
pipreqs                   # Библиотека для автоматического создания файла requirements.txt

requests==2.32.3          # Простая библиотека для отправки HTTP-запросов
aiohttp>=3.9.5            # Асинхронная HTTP-библиотека с улучшенной производительностью для работы с HTTP-запросами
openai==1.55.0            # Клиент для работы с API OpenAI

redis>=5.2.0              # Клиент для работы с Redis, подходящий для синхронных и асинхронных задач
SQLAlchemy>=2.0.36        # ORM (объектно-реляционное отображение) для работы с базами данных в Python
asyncpg>=0.30.0           # Асинхронный драйвер для работы с PostgreSQL базами данных
influxdb_client>=1.47.0   # Клиент для работы с InfluxDB

babel>=2.16.0             # Библиотека для интернационализации и локализации приложений